Pushing changes from partially abandoned project.
