+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "RESTy",
+  "name": "RESTy - A React-based API tester",
   "icons": [
     {
       "src": "favicon.ico",
