--- src/Footer/Footer.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import './Footer.scss';
-
-class Footer extends React.Component {
-  render() {
-    return (
-      <footer>
-        <p>&copy; 2021 - Jeremy Penning</p>
-      </footer>
-    )
-  }
-}
-
-export default Footer;
