+++ src/components/List/List.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class List extends React.Component{
+  render(){
+    return (
+      <ul>
+        {this.props.children}
+      </ul>
+    )
+  }
+}
+
+export default List;
