+++ src/index.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App/App';
+import App from './components/App/App';
+
+import './reset.scss';
+import './index.scss';
 
 ReactDOM.render(
   <React.StrictMode>
