+++ src/reset.scss
@@ -53,3 +53,25 @@ table {
 body, html {
   height: 100%;
 }
+
+/*
+    Button Reset from Chris Coyier - Adapted
+    https://css-tricks.com/overriding-default-button-styles/
+*/
+button {
+  display: inline-block;
+  border: none;
+  padding: 0;
+  margin: 0;
+  text-decoration: none;
+  background: white;
+  color: black;
+  font-family: sans-serif;
+  font-size: 1rem;
+  cursor: pointer;
+  text-align: center;
+  transition: background 250ms ease-in-out, 
+              transform 150ms ease;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+}
